+++ cdh-tsat6-stm32project-nucleo/.settings/com.st.stm32cube.ide.mcu.sfr.prefs
@@ -1,0 +1,3 @@
+eclipse.preferences.version=1
+svd_custom_file_path=
+svd_file_path=platform\:/plugin/com.st.stm32cube.ide.mcu.productdb.debug/resources/cmsis/STMicroelectronics_CMSIS_SVD/STM32L4x2.svd
