+++ cdh-tsat6-stm32project-nucleo/.settings/stm32cubeide.project.prefs
@@ -1,5 +1,6 @@
+2F62501ED4689FB349E356AB974DBE57=066877690AB3D91313F127A75C2A7618
 635E684B79701B039C64EA45C3F84D30=AC6501677675D8BDF426E7F5F1705774
 66BE74F758C12D739921AEA421D593D3=0
-8DF89ED150041C4CBC7CB9A9CAA90856=9DD1788FE739555F5A126365B7944BCB
-DC22A860405A8BF2F2C095E5B6529F12=9DD1788FE739555F5A126365B7944BCB
+8DF89ED150041C4CBC7CB9A9CAA90856=066877690AB3D91313F127A75C2A7618
+DC22A860405A8BF2F2C095E5B6529F12=0F49CA144B57946B10C42D9C67559D25
 eclipse.preferences.version=1
