+++ cdh-tsat6-stm32project-nucleo/Drivers/Hardware_Peripherals/Src/radio_spi_Si4464.c
@@ -7,7 +7,7 @@
  *      Nikolaus J. Reichert <nikolaus.reichert@umsats.ca>
  */
 #include <stdio.h>
-#include "Si446x/radio_spi_Si4464.h"
+#include "radio_spi_Si4464.h"
 
 
 /**
